refactored by separating endpoints
