+++ src/tendermint_rpc_canister/src/endpoints/mod.rs
@@ -1,0 +1,2 @@
+pub mod block;
+pub mod tx_sync;
